Removed unnecessary EntityListService
