+++ sdk-src/src/test/java/com/hpe/adm/nga/sdk/unit_tests/common/CommonMethods.java
@@ -108,7 +108,7 @@ public class CommonMethods {
 
         @Override
         public EntityList getEntityList(OctaneHttpClient octaneHttpClient, String baseDomain, String entityName) {
-            return new EntityList(octaneHttpClient, baseDomain, entityName);
+            return new EntityList(octaneHttpClient, baseDomain + entityName);
         }
     }
 }
